requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
discord.py==1.7.3
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
dpytest==0.5.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
iniconfig==1.1.1
multidict==5.1.0
packaging==21.0
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-django==4.4.0
pytz==2021.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
yarl==1.6.3
zope.interface==5.4.0
